Split off secp256k1 into its own file
